Set roles when authenticating via greenfield
fixes #1855
