+++ BTCPayServer/Security/GreenField/BasicAuthenticationHandler.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Security.Claims;
 using System.Text;
 using System.Text.Encodings.Web;
@@ -67,6 +68,7 @@ namespace BTCPayServer.Security.GreenField
                 new Claim(GreenFieldConstants.ClaimTypes.Permission,
                     Permission.Create(Policies.Unrestricted).ToString())
             };
+            claims.AddRange((await _userManager.GetRolesAsync(user)).Select(s => new Claim(_identityOptions.CurrentValue.ClaimsIdentity.RoleClaimType, s)));
 
             return AuthenticateResult.Success(new AuthenticationTicket(
                 new ClaimsPrincipal(new ClaimsIdentity(claims, GreenFieldConstants.AuthenticationType)),
